HistoBIT3D_Code/requirements.txt
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028896657/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.13.0
conda-build==3.21.9
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fonttools==4.57.0
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
guided-diffusion==0.0.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==2.10.1
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640704205123/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkginfo @ file:///tmp/build/80754af9/pkginfo_1643162084911/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2025.2
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
# Editable install with no version control (uvcgan2==0.0.1)
-e /home/durrlab/Desktop/Anthony/UGVSM/UVCGANv2_Original
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
